Keep shader graph properties order in particle output

+++ TestProjects/VisualEffectGraph_HDRP/Assets/AllTests/Editor/Tests/VFXTestCommon.cs
@@ -14,7 +14,7 @@ namespace UnityEditor.VFX.Test
 {
     class VFXTestCommon
     {
-        static readonly string tempBasePath = "Assets/TmpTests/";
+        public static readonly string tempBasePath = "Assets/TmpTests/";
         static readonly string tempFileFormat = tempBasePath + "vfx_{0}.vfx";
 
         public static VFXGraph CopyTemporaryGraph(string path)
@@ -45,14 +45,17 @@ namespace UnityEditor.VFX.Test
 
         public static void DeleteAllTemporaryGraph()
         {
-            foreach (string file in System.IO.Directory.GetFiles(tempBasePath))
+            if (Directory.Exists(tempBasePath))
             {
-                try
+                foreach (string file in System.IO.Directory.GetFiles(tempBasePath))
                 {
-                    AssetDatabase.DeleteAsset(file);
-                }
-                catch (System.Exception) // Don't stop if we fail to delete one asset
-                {
+                    try
+                    {
+                        AssetDatabase.DeleteAsset(file);
+                    }
+                    catch (System.Exception) // Don't stop if we fail to delete one asset
+                    {
+                    }
                 }
             }
         }
